examples/phi2-eo-tile-filter/requirements.txt
torch>=2.2
numpy>=1.26
onnx>=1.15
onnxruntime>=1.18
Pillow>=10.2
scikit-learn>=1.5
psutil>=5.9
pytest>=8.0
